fix typsense issues and update verionCreateUpdate function
